Created Intake
Created intake and intake command as well as joystick mapping based on SwerveTestBed. Because of Superstructure and RobotContainer, I did not use IntakeCommand (obsolete at the moment, may delete later). I ran commands in RobotContainer with a edge enhanced button controlling intake and outtake. These buttons HAD to be created in every type of controller config (kind of annoying) in order to add it as an enhanced button.