created climb time rumble feature
The manipulation controller will rumble a 1/4 of full power (same as driver controller's teleop init rumble) when the elapsed match time is equal to the time the drivers specifies on shuffleboard. The time from the driver station is a countdown, so the number the drivers enter must be how many seconds they want remaining before they climb. The check for the need to rumble is updated every .25 seconds to match the other subsystems. Wait command is in the method in order to stop the rumble after 2 seconds.